requirements.txt
# Core dependencies for LLM Fine-Tuning Platform - Phase 1
# Unsloth for fast fine-tuning
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git

# Core ML libraries
torch>=2.1.0
transformers>=4.36.0
datasets>=2.14.0
peft>=0.7.0
trl>=0.7.4
bitsandbytes>=0.41.0
accelerate>=0.25.0

# Data handling
pandas>=2.0.0
jsonlines>=4.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

# Phase 2: FastAPI Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Task Queue
celery>=5.3.0
redis>=5.0.0
flower>=2.0.1

# Storage (optional for S3)
boto3>=1.34.0
aiofiles>=23.2.1
